imp - Consolidated type-based info parsers
We've consolidated type-based info parsers

---

We've made a substantial change to type-based info parsers by determining the types from the value and parsing them more efficiently. We've cut down the constants so that there are no repeat types with a difference in the delimiter.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1