imp - Fixed out of bounds in some cases
We've fixed out of bounds when trying to split keys and values.

---

We've fixed out of bounds when trying to split keys and values in certain situations. This can be easily reproduced by importing a contact with a picture to Nitrocid KS, which internally uses the SaveTo() and Parse() functions for verification.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1