imp - brk|doc - Condensed parsers
---

We've condensed all the vCard parsers so that BaseVcardParser handles all the cases (vCard 2.1, 3.0, 4.0, and 5.0). This is necessary to make Card instances more secure.

This is a breaking change involving the card parsers. The only thing remaining is to get rid of an extra step to get all the cards. This way, we would have internalized the card parsers, which were initially reserved for internal use, and we would have made it one-step to getting an array of Cards.

---

Type: imp
Breaking: True
Doc Required: True
Part: 1/1