add - Added vCalendar specs
---

We've added just the specifications for vCalendar. However, we won't implement it in this version of VisualCard, because this version focuses on the core functionality only. However, in a future release, we may implement it in a limited form which will then get improved from time to time until we reach version 2.0 of VisualCard.

---

Type: add
Breaking: False
Doc Required: False
Part: 1/1