imp - prf - Reduce exceptions when parsing line
---

When PropertyInfo gets constructed with a possible multiline string or an
empty string, we need to avoid that in the CalendarTools and the CardTools
obtaining functions.

We're no longer relying on exceptions, so this means that VisualCard is
now back to speed!

---

Type: imp
Breaking: False
Doc Required: False
Backport Required: False
Part: 1/1